Prepare the next wave of datasets.

+++ charts/inputs/train/convert_from_labels.sh
@@ -12,11 +12,10 @@ input_dir="$1"
 output_dir="$2"
 
 if test -e "$output_dir"; then
-    echo "$output_dir already exists"
-    exit 1
+    echo "$output_dir already exists, not overwriting"
+else
+    cp -r "$input_dir" "$output_dir"
 fi
-
-cp -r "$input_dir" "$output_dir"
 
 cd "$output_dir"
 
@@ -24,23 +23,25 @@ echo "Generating aliased.png images.."
 for jf in *.json; do
     ((i=i%N)); ((i++==0)) && wait
     {
-        python3 "${HERE_DIR}/../generate_plots/generate_aliased_from_labels.py" "$jf" 
-        mv "$jf" "$jf".orig
+        if ! test -f "${jf%.json}.aliased.png"; then
+            python3 "${HERE_DIR}/../generate_plots/generate_aliased_from_labels.py" "$jf" 
+        fi
+        if ! test -f "$jf".orig; then
+            mv "$jf" "$jf".orig
+        fi
         echo -n '.'
     } &
 done
 wait
 
-echo
-echo "Renaming rendered.png -> antialiased.png.."
-rename 's/\.rendered.png/.antialiased.png/' *.rendered.png
-
 echo "Generating new json files.."
 for f in *.aliased.png; do
     ((i=i%N)); ((i++==0)) && wait
     {
-        "${HERE_DIR}/../generate_plots/gt_from_pairs/build/gt_from_pairs" "${f%.aliased.png}"
-        echo -n '.'
+        if ! test -f "${f%.aliased.png}.json"; then
+            "${HERE_DIR}/../generate_plots/gt_from_pairs/build/gt_from_pairs" "${f%.aliased.png}"
+        fi
+        echo -n '.'        
     } &
 done
 wait
